Testing-related fixes
